rag-backend/requirements.txt

fastapi
uvicorn
chromadb
sentence-transformers
PyMuPDF
python-docx
google-generativeai
